user/api/permissions.py
from rest_framework import permissions


class IsUserProfile(permissions.BasePermission):
    def has_object_permission(self, request, view, obj):
        if request.method in permissions.SAFE_METHODS:
            return True
        elif obj.email == request.user.email:
            return True
