requirements.txt
asgiref==3.5.0
black==22.1.0
boto3==1.21.26
botocore==1.24.26
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-allauth==0.49.0
django-appconf==1.0.5
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.0.0
django-imagekit==4.1.0
django-jazzmin==2.4.9
django-phonenumber-field==6.1.0
django-rest-auth==0.9.5
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
jmespath==1.0.0
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
phonenumbers==8.12.45
pilkit==2.0
Pillow==9.0.1
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
